Derive: Leverage new amrex::DeriveFuncMF
The addition of this new type of derive function in
AMReX-Codes/amrex#3990 means we can store the derive function to
calculate the relevant derived quantity in the derive_lst. This means we
can rely on AmrLevel::derive() for our needs and we can remove our
overrides of this function in GRAMRLevel and BinaryBHLevel.

The Weyl4 and Constaints tests have been modified to use the new
DeriveFuncMF functions.