Build: Remove default TMP_BUILD_DIR
I think this will help with #68. The disadvantage (and the reason I
bothered to set this default in the first place) is that consecutive
builds of different AMReX executables (e.g. two different examples) will
be slower. To mitigate this, users should use Ccache. I will add this to
the CI in a separate PR.