Ported velocity and normalisation/delta function
1. Ported normalisation (2pi/L)^3 and delta function/mult. by -1/2
for position and velocity field.
2. Checked the normalisation and everything against GRChombo with no
window or randomisation.
3. Added memory allocation for Aij field. Juliana said doing both hij and
Aij assignment in the same for loop can be faster if you're not operating
at the edge of the available RAM, which is at least true for cosmosx. So
I'm starting with that strategy.
4. Adjusted the randomisation method so that the random draws are made at
each polarisation step in the box loop, and fed into the random_field function
along with the spectrum type, so that each polarisation's hij and Aij get the
same draw, and each polarisation gets a different draw.
5. Followed through the Aij port. Now all that's left is to put everything on the grid.