Refactor tensor extraction routines out of RandomField
This commit takes the extraction and derive routines out of RandomFields
and puts them in a child class, TensorExtraction. It also moves the
RandomField class from Source/InitialConditions to Examples/ScalarField.
The new TensorExtraction files are also in Examples/ScalarField.

This refactor involved altering the calls to the extraction and derive
functions in ScalarFieldLevel.cpp, and altering the loading of variables
in SimulationParameters.hpp.

Within RandomField, I took L out of the params struct and made it a
class data member, so it could be accessed by TensorExtraction. I also
changed some of the inheretance permissions in RandomField so that
TensorExtraction could access all of the necessary functions,
particularly for the FFT.

I also fixed a warning relating to the lack of return in
RandomField::calculate_basis_vector in the final logic statement, and I
updated the Make.packages in Source/InitialConditions and
Examples/ScalarField according to the above changes.